fix(benchmarks): aplica correcoes do cycle-review (2 HIGH + MEDIUMs)
Review de 4 especialistas (arquitetura/testes/cross-validation/dominio): 0 BLOCKER.
- HIGH: CHANGELOG sincronizado com o JSON de evidencia (recall exato, QPS faixa).
- HIGH: testes reais de recall distance-vs-identity + fronteira do eps (eram prose-only).
- MED (dominio): ground-truth arredonda para float32 (casa com o storage vector(float4) do pgvector).
- MED (arquitetura): db._cursor tipa erros operacionais psycopg2 -> DBUnavailableError; docstring honesto.
- MED (dominio): warmup nao-cronometrado; relatorio auto-descritivo (host/methodology/n_queries).
- MED: remove metrica 'ip' meio-wired; testes MissVectorDB (recall 0.0) + round-trip do relatorio.

42 testes verdes (era 36), coverage 98% (recall/metrics/dataset 100%), ruff+vulture limpos.
Evidencia atualizada: HNSW recall@10 0.841/0.960 (ef_search 40/100), seed=42 n=5000 dim=128.

+++ benchmarks/theodb_bench/recall.py
@@ -16,8 +16,11 @@ _METRICS = ("l2", "cosine")
 
 
 def _pairwise_distances(corpus: np.ndarray, queries: np.ndarray, metric: str) -> np.ndarray:
-    corpus = np.asarray(corpus, dtype=np.float64)
-    queries = np.asarray(queries, dtype=np.float64)
+    # pgvector stores `vector` as float4 (single precision); the exact oracle must rank the SAME
+    # float32-precision values the index sees, else near-ties diverge between oracle and SUT.
+    # Round to float32, then compute the distance itself in float64 for numerical stability. (review M1)
+    corpus = np.asarray(corpus, dtype=np.float32).astype(np.float64)
+    queries = np.asarray(queries, dtype=np.float32).astype(np.float64)
     if corpus.shape[0] == 0:
         raise ValueError("corpus is empty")
     if queries.shape[0] == 0:
